pubmed-affiliation-filter/pyproject.toml
[tool.poetry]
name = "pubmed_affiliation_filter"
version = "0.1.0"
description = "Filter PubMed papers by author affiliations with pharmaceutical and biotech companies"
authors = ["Onepice28 <144516580+Onepice28@users.noreply.github.com>"]
readme = "README.md"
packages = [
    { include = "pubmed_affiliation_filter", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
requests = ">=2.32.0,<3.0.0"
biopython = ">=1.83,<2.0.0"
pandas = ">=2.2.0,<3.0.0"
typing-extensions = ">=4.9.0,<5.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
mypy = "^1.7.1"

[tool.poetry.scripts]
get-papers-list = "pubmed_affiliation_filter.cli:main"

